feat(OPAL-158): checking if the deletable kalendar belongs to the user

+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/controllers/KalendarController.java
@@ -68,9 +68,12 @@ public class KalendarController {
     @Transactional
     @DeleteMapping(value = "/{id}/deletecalendar")
     public ResponseEntity deleteCalendar(@RequestHeader("X-Client-Token") String clientToken,
-                                         @PathVariable(name = "id") long id,
-                                         HttpServletRequest request) {
-        kalendarService.deleteKalendarAndGoogleCalendarById(id);
-        return new ResponseEntity(HttpStatus.OK);
+                                         @PathVariable(name = "id") long id) {
+        if (kalendarService.theKalendarBelongsToTheUser(clientToken, id)) {
+            kalendarService.deleteKalendarAndGoogleCalendarById(id);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("This Kalendar doesn't belongs to the current user", HttpStatus.UNAUTHORIZED);
+        }
     }
 }
